Add RETGUARD to clang for amd64 and arm64.
This security mechanism uses per-function random cookies to protect access
to function return instructions, with the effect that the integrity of the
return address is protected and function return instructions are harder
to use in ROP gadgets.

On function entry the return address is combined with a per-function random
cookie and stored in the stack frame. The integrity of this value is verified
before function return, and if this check fails, the program aborts. In this way
RETGUARD is an improved stack protector, since the cookies are per-function. The
verification routine is constructed such that the binary space immediately
before each ret instruction is padded with trap instructions, which makes these
return instructions difficult to use in ROP gadgets.

Remaining work includes adding this mechanism to assembly routines, which must
be done by hand. Many thanks to all those who helped test and provide feedback,
especially deaadt, tb, espie and naddy.

ok deraadt@