Fix -msave-args epilogue insertion on amd64.
Also fixes frame lowering in general.

The use of LEAVE in the -msave-args case previously was implemented
by moving the terminating frame pointer restore to the end of x86
epilogue insertion. With save-args this is a LEAVE, and otherwise is
a POP $rbp. The move was to ensure that the rest of frame lowering
could iterate back over register restores to find the right place
to insert stack adjustment, and since these all use POP $reg, the frame
pointer restore did not cause this to stop in the wrong spot. Moving
the frame pointer restore to the end of the epilogue insertion broke
other things in clang 11 though, so now we insert the LEAVE in the
same spot as upstream, but also ensure that we can back up over LEAVE
in addition to POP instructions.