Shuffle how lldb register contexts are built.
Instead of using #if defined(__arch__) to include / exclude the entire contents
of the NativeRegisterContext implementations, use a single NativeRegisterContextOpenBSD_arch
which includes the right arch specific register context, and provides a dummy implementation
for unsupported architectures.

This allows building lldb on architectures which do not have a register context implementation
so it can be used as a remote client.

ok patrick@